try to make shutdown logic more straightforward
also, move port into it's own file. get rid of unused test code
